OPMS_v3-dev3.1/OPMS_v3-dev3.1/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.3
Django==2.0.6
django-pure-pagination==0.3.0
idna==2.7
paramiko==2.4.1
Pillow~=9.5.0
pyasn1==0.4.3
pycparser==2.18
PyMySQL==0.8.1
PyNaCl==1.2.1
pytz==2018.4
six==1.11.0
tornado==5.1
xlwt==1.3.0
django-simple-captcha==0.5.11
PyJWT~=1.7.1
requests~=2.30.0
channels~=4.0.0